requirements.txt
fastapi
uvicorn
requests
pyyaml
psutil
